+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All Notable changes to `Slytherin` will be documented in this file
 ## [0.4.0](https://github.com/rougin/slytherin/compare/v0.3.0...v0.4.0)
 
 ### Added
-- `ComponentCollection::setContainer` for adding libraries that is implemented in [container-interop/container-interop](https://github.com/container-interop/container-interop)'s `ContainerInterface`
+- `ComponentCollection::setContainer` for adding packages implemented on [container-interop/container-interop](https://github.com/container-interop/container-interop)'s `ContainerInterface`
 
 ### Changed
 - `ErrorHandler` to `Debug`
@@ -13,8 +13,8 @@ All Notable changes to `Slytherin` will be documented in this file
 - Renamed sample library implementations
 
 ### Removed
-- `Http` directory (Will now adopt [PSR-7](http://www.php-fig.org/psr/psr-7/) implementations)
 - `ComponentCollection::setInjector`
+- `Http` directory (Will now require implementations in [PSR-7](http://www.php-fig.org/psr/psr-7))
 
 ## [0.3.0](https://github.com/rougin/slytherin/compare/v0.2.1...v0.3.0) - 2015-11-02